// pdf-editor/src/code/contractManagement/builder/types.ts
export interface ImageElement extends BlockStyle {
  type: 'image';
  id: string;
  order: number; 
  height: number;
  width?: number; 
  imageData?: string; 
  imageUrl?: string;
  page: number;
  align?: 'left' | 'center' | 'right';
  imageEffect?: 'none' | 'grayscale';
}
// ...
export interface BoxSpacing {
  top: number;
  right: number;
  bottom: number;
  left: number;
}

export interface BlockStyle {
  backgroundColor?: string;
  padding?: BoxSpacing;
  margin?: BoxSpacing;
}